requirements.txt
docopt
schedule
RPi.GPIO
